KAFKA-646 Squashed MapR specific commits (from https://github.com/mapr/private-kafka/commit/1bf544c5054798fb47210e407362d1c5c2d5c3de to https://github.com/mapr/private-kafka/commit/a25b751d2ceee1bc0b777d11d166f14cd8c20b3d) are copied from branch 2.1.1-mapr (https://github.com/mapr/private-kafka/tree/2.1.1-mapr) to branch branch-2.6.1-mapr

+++ clients/src/main/java/org/apache/kafka/clients/consumer/internals/ConsumerCoordinator.java
@@ -93,7 +93,7 @@ public final class ConsumerCoordinator extends AbstractCoordinator {
     private final SubscriptionState subscriptions;
     private final OffsetCommitCallback defaultOffsetCommitCallback;
     private final boolean autoCommitEnabled;
-    private final int autoCommitIntervalMs;
+    private final long autoCommitIntervalMs;
     private final ConsumerInterceptors<?, ?> interceptors;
     private final AtomicInteger pendingAsyncCommits;
 
@@ -146,7 +146,7 @@ public final class ConsumerCoordinator extends AbstractCoordinator {
                                String metricGrpPrefix,
                                Time time,
                                boolean autoCommitEnabled,
-                               int autoCommitIntervalMs,
+                               long autoCommitIntervalMs,
                                ConsumerInterceptors<?, ?> interceptors,
                                boolean throwOnFetchStableOffsetsUnsupported) {
         super(rebalanceConfig,
